+++ changelog.md
@@ -1,3 +1,11 @@
+### v0.9.0-alpha2
+
+* 2018/5/21
+* Remove API LibDeflate:VerifyDictionary
+* Remove API LibDeflate:DictForWoW
+* Changed API LibDeflate:CreateDictionary
+
 ### v0.9.0-alpha1
-2018/5/21
-The first working version.
+
+* 2018/5/20
+* The first working version.